initialize cobra
