+++ go.mod
@@ -1,5 +1,10 @@
 module github.com/aifaniyi/bootstrap
 
-go 1.12
+go 1.15
 
-require github.com/iancoleman/strcase v0.1.3
+require (
+	github.com/iancoleman/strcase v0.1.3
+	github.com/mitchellh/go-homedir v1.1.0
+	github.com/spf13/cobra v1.1.3
+	github.com/spf13/viper v1.7.1
+)
